Port hardik-master d987910: fix MO-footprint score saturation bug
skeleton_graph_mo_footprint_calc used multiplicative scoring
(score *= (1.0 + ratio) per MO edge), which scales exponentially with
graph size -- any graph with >= 7 writes would mathematically saturate
to the 100.0 cap regardless of whether its edges were actually rare or
common. Every real benchmark tested this session (barrier, sb-loop)
has well beyond 7 writes, so this bug affected essentially all
non-trivial fuzzing runs.

Replaced with averaging: each edge's rarity ratio is now clamped to
[1.0, 100.0] individually, then the function returns the mean across
all MO edges in the graph. Return type changed u32 -> double to match
(the previous u32 truncation was itself a precision bug, separate from
the saturation issue).

Verified against barrier: mo coverage climbed to 10 (vs the typical
7-9 seen before) with genuinely varied edge frequencies (1 to 585
times) in one 20s run, consistent with scores now differentiating
rare vs common edges instead of all saturating together.

+++ FUZZER_Rebuilt/Main/include/mo-footprint.hpp
@@ -9,7 +9,7 @@ extern "C"{
 #include "skeleton_graph.hpp"
 #include "skeleton_graph_events.hpp"
 
-u32 skeleton_graph_mo_footprint_calc(SkeletonGraph* graph);
+double skeleton_graph_mo_footprint_calc(SkeletonGraph* graph);
 
 
 #endif
